+++ wgsl_to_wgpu/src/compute.rs
@@ -14,7 +14,7 @@ where
         .filter_map(|e| {
             if e.stage == naga::ShaderStage::Compute {
                 let workgroup_size_constant = workgroup_size(e, demangle.clone());
-                let create_pipeline = create_compute_pipeline(e, demangle.clone());
+                let create_pipeline = create_compute_pipeline(module, e, demangle.clone());
 
                 Some(quote! {
                     #workgroup_size_constant
@@ -38,7 +38,11 @@ where
     }
 }
 
-fn create_compute_pipeline<F>(e: &naga::EntryPoint, demangle: F) -> TokenStream
+fn create_compute_pipeline<F>(
+    module: &naga::Module,
+    e: &naga::EntryPoint,
+    demangle: F,
+) -> TokenStream
 where
     F: Fn(&str) -> TypePath,
 {
@@ -52,18 +56,39 @@ where
 
     // TODO: Include a user supplied module name in the label?
     let label = format!("Compute Pipeline {name}");
-    quote! {
-        pub fn #pipeline_name(device: &wgpu::Device) -> wgpu::ComputePipeline {
-            let module = super::create_shader_module(device);
-            let layout = super::create_pipeline_layout(device);
-            device.create_compute_pipeline(&wgpu::ComputePipelineDescriptor {
-                label: Some(#label),
-                layout: Some(&layout),
-                module: &module,
-                entry_point: Some(#entry_point),
-                compilation_options: Default::default(),
-                cache: Default::default(),
-            })
+
+    if !module.overrides.is_empty() {
+        quote! {
+            pub fn #pipeline_name(device: &wgpu::Device, overrides: &super::OverrideConstants) -> wgpu::ComputePipeline {
+                let module = super::create_shader_module(device);
+                let layout = super::create_pipeline_layout(device);
+                device.create_compute_pipeline(&wgpu::ComputePipelineDescriptor {
+                    label: Some(#label),
+                    layout: Some(&layout),
+                    module: &module,
+                    entry_point: Some(#entry_point),
+                    compilation_options: wgpu::PipelineCompilationOptions {
+                        constants: &overrides.constants(),
+                        ..Default::default()
+                    },
+                    cache: Default::default(),
+                })
+            }
+        }
+    } else {
+        quote! {
+            pub fn #pipeline_name(device: &wgpu::Device) -> wgpu::ComputePipeline {
+                let module = super::create_shader_module(device);
+                let layout = super::create_pipeline_layout(device);
+                device.create_compute_pipeline(&wgpu::ComputePipelineDescriptor {
+                    label: Some(#label),
+                    layout: Some(&layout),
+                    module: &module,
+                    entry_point: Some(#entry_point),
+                    compilation_options: Default::default(),
+                    cache: Default::default(),
+                })
+            }
         }
     }
 }