+++ wgsl_to_wgpu/src/lib.rs
@@ -816,6 +816,10 @@ mod test {
                     let x = h;
                 }
             }
+
+            @compute
+            @workgroup_size(256)
+            fn main2() {}
         "#};
 
         let actual = create_shader_module(source, "shader.wgsl", WriteOptions::default()).unwrap();
@@ -833,6 +837,57 @@ mod test {
 
             @fragment
             fn fs_multiple() -> Output {}
+        "#};
+        let actual = create_shader_module(source, "shader.wgsl", WriteOptions::default()).unwrap();
+        assert_rust_snapshot!(actual);
+    }
+
+    #[test]
+    fn create_shader_module_compute_overrides() {
+        let source = indoc! {r#"
+            struct Uniforms { 
+                color_rgb: vec3<f32>,
+            }
+
+            @group(0) @binding(0)
+            var<storage, read_write> uniforms: Uniforms;
+
+            override dt: f32;
+
+            @compute
+            @workgroup_size(1)
+            fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
+                if global_id.x == 0u {
+                    uniforms.color_rgb = vec3(1.0+dt);
+                }
+            }
+        "#};
+        let actual = create_shader_module(source, "shader.wgsl", WriteOptions::default()).unwrap();
+        assert_rust_snapshot!(actual);
+    }
+
+    #[test]
+    fn create_shader_module_overrides() {
+        let source = indoc! {r#"
+            override b1: bool = true;
+            override b2: bool = false;
+            override b3: bool;
+
+            override f1: f32 = 0.5;
+            override f2: f32;
+
+            override f3: f64 = 0.6;
+            override f4: f64;
+
+            override i1: i32 = 0;
+            override i2: i32;
+            override i3: i32 = i1 * i2;
+
+            @id(0) override a: f32 = 1.0;
+            @id(35) override b: f32 = 2.0;
+
+            @fragment
+            fn main() {}
         "#};
         let actual = create_shader_module(source, "shader.wgsl", WriteOptions::default()).unwrap();
         assert_rust_snapshot!(actual);
@@ -1039,38 +1094,6 @@ mod test {
         let actual = vertex_struct_methods(&module, demangle_identity);
 
         assert_tokens_snapshot!(items_to_tokens(actual));
-    }
-
-    #[test]
-    fn write_compute_module_empty() {
-        let source = indoc! {r#"
-            @vertex
-            fn main() {}
-        "#};
-
-        let module = naga::front::wgsl::parse_str(source).unwrap();
-        let actual = compute_module(&module, demangle_identity);
-
-        assert_tokens_eq!(quote!(), actual);
-    }
-
-    #[test]
-    fn write_compute_module_multiple_entries() {
-        let source = indoc! {r#"
-            @compute
-            @workgroup_size(1,2,3)
-            fn main1() {}
-
-            @compute
-            @workgroup_size(256)
-            fn main2() {}
-        "#
-        };
-
-        let module = naga::front::wgsl::parse_str(source).unwrap();
-        let actual = compute_module(&module, demangle_identity);
-
-        assert_tokens_snapshot!(actual);
     }
 
     #[test]