+++ wgsl_to_wgpu/src/lib.rs
@@ -58,12 +58,15 @@ use syn::Ident;
 use thiserror::Error;
 
 mod bindgroup;
+mod compute;
 mod consts;
 mod entry;
 mod structs;
 mod wgsl;
 
 pub use naga::valid::Capabilities as WgslCapabilities;
+
+use crate::compute::compute_module;
 
 /// Errors while generating Rust source for a WGSL shader module.
 #[derive(Debug, Error)]
@@ -667,86 +670,6 @@ fn indexed_name_to_ident(name: &str, index: u32) -> Ident {
     Ident::new(&format!("{name}{index}"), Span::call_site())
 }
 
-fn compute_module<F>(module: &naga::Module, demangle: F) -> TokenStream
-where
-    F: Fn(&str) -> TypePath + Clone,
-{
-    let entry_points: Vec<_> = module
-        .entry_points
-        .iter()
-        .filter_map(|e| {
-            if e.stage == naga::ShaderStage::Compute {
-                let workgroup_size_constant = workgroup_size(e, demangle.clone());
-                let create_pipeline = create_compute_pipeline(e, demangle.clone());
-
-                Some(quote! {
-                    #workgroup_size_constant
-                    #create_pipeline
-                })
-            } else {
-                None
-            }
-        })
-        .collect();
-
-    if entry_points.is_empty() {
-        // Don't include empty modules.
-        quote!()
-    } else {
-        quote! {
-            pub mod compute {
-                #(#entry_points)*
-            }
-        }
-    }
-}
-
-fn create_compute_pipeline<F>(e: &naga::EntryPoint, demangle: F) -> TokenStream
-where
-    F: Fn(&str) -> TypePath,
-{
-    let name = &demangle(&e.name).name;
-
-    // Compute pipeline creation has few parameters and can be generated.
-    let pipeline_name = Ident::new(&format!("create_{name}_pipeline"), Span::call_site());
-
-    // The entry name string itself should remain mangled to match the WGSL code.
-    let entry_point = &e.name;
-
-    // TODO: Include a user supplied module name in the label?
-    let label = format!("Compute Pipeline {name}");
-    quote! {
-        pub fn #pipeline_name(device: &wgpu::Device) -> wgpu::ComputePipeline {
-            let module = super::create_shader_module(device);
-            let layout = super::create_pipeline_layout(device);
-            device.create_compute_pipeline(&wgpu::ComputePipelineDescriptor {
-                label: Some(#label),
-                layout: Some(&layout),
-                module: &module,
-                entry_point: Some(#entry_point),
-                compilation_options: Default::default(),
-                cache: Default::default(),
-            })
-        }
-    }
-}
-
-fn workgroup_size<F>(e: &naga::EntryPoint, demangle: F) -> TokenStream
-where
-    F: Fn(&str) -> TypePath + Clone,
-{
-    let name = &demangle(&e.name).name;
-
-    let name = Ident::new(
-        &format!("{}_WORKGROUP_SIZE", name.to_uppercase()),
-        Span::call_site(),
-    );
-    let [x, y, z] = e
-        .workgroup_size
-        .map(|s| Literal::usize_unsuffixed(s as usize));
-    quote!(pub const #name: [u32; 3] = [#x, #y, #z];)
-}
-
 fn quote_shader_stages(stages: wgpu::ShaderStages) -> TokenStream {
     if stages == wgpu::ShaderStages::all() {
         quote!(wgpu::ShaderStages::all())
@@ -893,6 +816,10 @@ mod test {
                     let x = h;
                 }
             }
+
+            @compute
+            @workgroup_size(256)
+            fn main2() {}
         "#};
 
         let actual = create_shader_module(source, "shader.wgsl", WriteOptions::default()).unwrap();
@@ -910,6 +837,57 @@ mod test {
 
             @fragment
             fn fs_multiple() -> Output {}
+        "#};
+        let actual = create_shader_module(source, "shader.wgsl", WriteOptions::default()).unwrap();
+        assert_rust_snapshot!(actual);
+    }
+
+    #[test]
+    fn create_shader_module_compute_overrides() {
+        let source = indoc! {r#"
+            struct Uniforms { 
+                color_rgb: vec3<f32>,
+            }
+
+            @group(0) @binding(0)
+            var<storage, read_write> uniforms: Uniforms;
+
+            override dt: f32;
+
+            @compute
+            @workgroup_size(1)
+            fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
+                if global_id.x == 0u {
+                    uniforms.color_rgb = vec3(1.0+dt);
+                }
+            }
+        "#};
+        let actual = create_shader_module(source, "shader.wgsl", WriteOptions::default()).unwrap();
+        assert_rust_snapshot!(actual);
+    }
+
+    #[test]
+    fn create_shader_module_overrides() {
+        let source = indoc! {r#"
+            override b1: bool = true;
+            override b2: bool = false;
+            override b3: bool;
+
+            override f1: f32 = 0.5;
+            override f2: f32;
+
+            override f3: f64 = 0.6;
+            override f4: f64;
+
+            override i1: i32 = 0;
+            override i2: i32;
+            override i3: i32 = i1 * i2;
+
+            @id(0) override a: f32 = 1.0;
+            @id(35) override b: f32 = 2.0;
+
+            @fragment
+            fn main() {}
         "#};
         let actual = create_shader_module(source, "shader.wgsl", WriteOptions::default()).unwrap();
         assert_rust_snapshot!(actual);
@@ -1116,38 +1094,6 @@ mod test {
         let actual = vertex_struct_methods(&module, demangle_identity);
 
         assert_tokens_snapshot!(items_to_tokens(actual));
-    }
-
-    #[test]
-    fn write_compute_module_empty() {
-        let source = indoc! {r#"
-            @vertex
-            fn main() {}
-        "#};
-
-        let module = naga::front::wgsl::parse_str(source).unwrap();
-        let actual = compute_module(&module, demangle_identity);
-
-        assert_tokens_eq!(quote!(), actual);
-    }
-
-    #[test]
-    fn write_compute_module_multiple_entries() {
-        let source = indoc! {r#"
-            @compute
-            @workgroup_size(1,2,3)
-            fn main1() {}
-
-            @compute
-            @workgroup_size(256)
-            fn main2() {}
-        "#
-        };
-
-        let module = naga::front::wgsl::parse_str(source).unwrap();
-        let actual = compute_module(&module, demangle_identity);
-
-        assert_tokens_snapshot!(actual);
     }
 
     #[test]