+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unrelesed
+### Added
+* Added support for pipeline-overrideable constants in generated code for compute shaders.
+
 ### Changed
 * Updated to encase 0.12 from 0.11